+++ apps/server/src/schema/integrations/integration-types.ts
@@ -7,7 +7,7 @@ import {
   prisma,
   PublisherEnum,
 } from '@repo/database';
-import { MetaError } from '@repo/channel-utils';
+import { type FilterInsightsInputType, insightsColumnsGroupBy, insightsColumnsOrderBy, MetaError, type InsightsSearchExpression, InsightsSearchField, InsightsSearchOperator, type InsightsSearchTerm, } from '@repo/channel-utils';
 import { getDateDiffIn, getTomorrowStartOfDay, type IntervalType } from '@repo/utils';
 import type { InputShapeFromFields } from '@pothos/core';
 import { getRootOrganizationId } from '@repo/organization';
@@ -15,12 +15,6 @@ import { builder } from '../builder';
 import { ErrorInterface } from '../errors';
 import { type ChannelInitialProgressPayload } from '../pubsub';
 import type { GraphQLContext } from '../../context';
-import {
-  type InsightsSearchExpression,
-  InsightsSearchField,
-  InsightsSearchOperator,
-  type InsightsSearchTerm,
-} from '../../contexts/insights';
 
 export enum IntegrationStatusEnum {
   ComingSoon = 'ComingSoon',
@@ -206,28 +200,10 @@ export const AdAccountDto = builder.prismaObject('AdAccount', {
   }),
 });
 
-const insightsColumnsOrderBy = [
-  'spend_abs',
-  'impressions_abs',
-  'cpm_abs',
-  'spend_rel',
-  'impressions_rel',
-  'cpm_rel',
-] as const;
 export const InsightsColumnsOrderByDto = builder.enumType('InsightsColumnsOrderBy', {
   values: insightsColumnsOrderBy,
 });
 
-const insightsColumnsGroupBy = [
-  'adAccountId',
-  'adId',
-  'adSetId',
-  'campaignId',
-  'device',
-  'position',
-  'publisher',
-] as const;
-export type InsightsColumnsGroupByType = (typeof insightsColumnsGroupBy)[number];
 export const InsightsColumnsGroupByDto = builder.enumType('InsightsColumnsGroupBy', {
   values: insightsColumnsGroupBy,
 });
@@ -354,52 +330,53 @@ export const InsightsSearchExpressionDto = builder
     }),
   });
 
-export const FilterInsightsInput = builder.inputType('FilterInsightsInput', {
-  fields: (t) => ({
-    adAccountIds: t.stringList({ required: false }),
-    adIds: t.stringList({ required: false }),
-    dateFrom: t.field({ type: 'Date', required: false }),
-    dateTo: t.field({ type: 'Date', required: false }),
-    devices: t.field({ type: [DeviceEnumDto], required: false }),
-    groupBy: t.field({ type: [InsightsColumnsGroupByDto], required: false }),
-    interval: t.field({ type: InsightsIntervalDto, required: true }),
-    order: t.field({ type: OrderByDto, defaultValue: 'desc' }),
-    orderBy: t.field({ type: InsightsColumnsOrderByDto, required: true, defaultValue: 'cpm_rel' }),
-    page: t.int({
-      required: true,
-      description: 'Starting at 1',
-      defaultValue: 1,
-      validate: { min: [1, { message: 'Minimum page is 1' }] },
+  export const FilterInsightsInputDto = builder
+  .inputRef<FilterInsightsInputType>('FilterInsightsInput')
+  .implement({
+    fields: (t) => ({
+      adAccountIds: t.stringList({ required: false }),
+    adIds: t.stringList({ required: false }),dateFrom: t.field({ type: 'Date', required: false }),
+      dateTo: t.field({ type: 'Date', required: false }),
+      devices: t.field({ type: [DeviceEnumDto], required: false }),
+      groupBy: t.field({ type: [InsightsColumnsGroupByDto], required: false }),
+      interval: t.field({ type: InsightsIntervalDto, required: true }),
+      order: t.field({ type: OrderByDto, defaultValue: 'desc' }),
+      orderBy: t.field({ type: InsightsColumnsOrderByDto, required: true, defaultValue: 'cpm_rel' }),
+      page: t.int({
+        required: true,
+        description: 'Starting at 1',
+        defaultValue: 1,
+        validate: { min: [1, { message: 'Minimum page is 1' }] },
+      }),
+      pageSize: t.int({
+        required: true,
+        defaultValue: 12,
+        validate: {
+          min: [1, { message: 'Minimum page is 1' }],
+          max: [100, { message: 'Page size should not be more than 100' }],
+        },
+      }),
+      positions: t.field({ type: [InsightsPositionDto], required: false }),
+      publishers: t.field({ type: [PublisherEnumDto], required: false }),
+      search: t.field({ type: InsightsSearchExpressionDto, required: false }),
     }),
-    pageSize: t.int({
-      required: true,
-      defaultValue: 12,
-      validate: {
-        min: [1, { message: 'Minimum page is 1' }],
-        max: [100, { message: 'Page size should not be more than 100' }],
-      },
-    }),
-    positions: t.field({ type: [InsightsPositionDto], required: false }),
-    publishers: t.field({ type: [PublisherEnumDto], required: false }),
-    search: t.field({ type: InsightsSearchExpressionDto, required: false }),
-  }),
-  validate: [
-    [
-      (args) => !(args.dateFrom && args.dateTo && args.dateFrom.getTime() > args.dateTo.getTime()),
-      { message: 'Date from should be less or equal to date to' },
+    validate: [
+      [
+        (args) => !(args.dateFrom && args.dateTo && args.dateFrom.getTime() > args.dateTo.getTime()),
+        { message: 'Date from should be less or equal to date to' },
+      ],
+      [
+        (args) =>
+          !(
+            args.dateFrom &&
+            args.dateTo &&
+            args.interval === 'day' &&
+            args.dateTo.getTime() - args.dateFrom.getTime() > 1000 * 60 * 60 * 24 * 90
+          ),
+        { message: 'Day intervals cannot be more than 90 days' },
+      ],
     ],
-    [
-      (args) =>
-        !(
-          args.dateFrom &&
-          args.dateTo &&
-          args.interval === 'day' &&
-          args.dateTo.getTime() - args.dateFrom.getTime() > 1000 * 60 * 60 * 24 * 90
-        ),
-      { message: 'Day intervals cannot be more than 90 days' },
-    ],
-  ],
-});
+  });
 
 export const InsightsDatapointsInput = builder.inputType('InsightsDatapointsInput', {
   fields: (t) => ({
@@ -416,7 +393,6 @@ export const InsightsDatapointsInput = builder.inputType('InsightsDatapointsInpu
   }),
 });
 
-export type FilterInsightsInputType = typeof FilterInsightsInput.$inferInput;
 export type InsightsDatapointsInputType = typeof InsightsDatapointsInput.$inferInput;
 
 export const getIntegrationStatus = (integration: Integration | undefined): IntegrationStatusEnum => {
