+++ packages/channel-utils/src/insights-utils.ts
@@ -1,4 +1,4 @@
-import { type AdAccount, type Integration, Prisma, prisma } from '@repo/database';
+import { type AdAccount, type CurrencyEnum, type DeviceEnum, type Integration, Prisma, prisma, type PublisherEnum } from '@repo/database';
 import { logger } from '@repo/logger';
 import { AError, isAError } from '@repo/utils';
 import _ from 'lodash';
@@ -222,3 +222,97 @@ export const saveInsightsAdsAdsSetsCampaigns = async (
 
   await saveInsights(insights, adExternalIdMap, adAccount);
 };
+
+export interface GroupedInsightsWithEdges {
+  hasNext: boolean;
+  page: number;
+  pageSize: number;
+  edges: GroupedInsightWithDetails[];
+}
+
+export interface GroupedInsightWithDetails {
+  id: string;
+  adId?: string | null;
+  adAccountId?: string | null;
+  adSetId?: string | null;
+  campaignId?: string | null;
+  currency: CurrencyEnum;
+  device?: DeviceEnum | null;
+  publisher?: PublisherEnum | null;
+  position?: string | null;
+  datapoints: Datapoints[];
+}
+
+export interface Datapoints {
+  spend: bigint;
+  impressions: bigint;
+  cpm?: bigint | null;
+  date: Date;
+}
+
+export const insightsColumnsOrderBy = [
+  'spend_abs',
+  'impressions_abs',
+  'cpm_abs',
+  'spend_rel',
+  'impressions_rel',
+  'cpm_rel',
+] as const;
+
+export type InsightsColumnsOrderByType = (typeof insightsColumnsOrderBy)[number];
+  export type InsightsPositionType = 'an_classic' | 'biz_disco_feed' | 'facebook_reels' | 'facebook_reels_overlay' | 'facebook_stories' | 
+'feed' | 'instagram_explore' | 'instagram_explore_grid_home' | 'instagram_profile_feed' | 'instagram_reels' | 
+'instagram_search' | 'instagram_stories' | 'instream_video' | 'marketplace' | 'messenger_inbox' | 
+'messenger_stories' | 'rewarded_video' | 'right_hand_column' | 'search' | 'video_feeds' | 'unknown';
+
+export const insightsColumnsGroupBy = [
+  'adAccountId',
+  'adId',
+  'adSetId',
+  'campaignId',
+  'device',
+  'position',
+  'publisher',
+] as const;
+export type InsightsColumnsGroupByType = (typeof insightsColumnsGroupBy)[number];
+
+export enum InsightsSearchField {
+  AdName = 'a.name',
+  AccountName = 'aa.name',
+  AdSetName = 'ase.name',
+  CampaignName = 'c.name',
+}
+export enum InsightsSearchOperator {
+  Contains = 'contains',
+  StartsWith = 'startsWith',
+  Equals = 'equals',
+}
+
+export interface InsightsSearchTerm {
+  field: InsightsSearchField;
+  operator: InsightsSearchOperator;
+  value: string;
+}
+
+export interface InsightsSearchExpression {
+  and?: InsightsSearchExpression[] | null;
+  or?: InsightsSearchExpression[] | null;
+  term?: InsightsSearchTerm | null;
+}
+
+
+
+export interface FilterInsightsInputType {
+  dateFrom?: Date | null;
+  dateTo?: Date | null;
+  devices?: DeviceEnum[] | null;
+  groupBy?: InsightsColumnsGroupByType[] | null;
+  interval: 'day' | 'week' | 'month' | 'quarter';
+  order?: 'asc' | 'desc' | null;
+  orderBy: InsightsColumnsOrderByType;
+  page: number;
+  pageSize: number;
+  positions?: InsightsPositionType[] | null;
+  publishers?: PublisherEnum[] | null;
+  search?: InsightsSearchExpression | null;
+}
